% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{print.qswat_project}
\alias{print.qswat_project}
\title{Print QSWATPlus Project Summary}
\usage{
\method{print}{qswat_project}(x, ...)
}
\arguments{
\item{x}{A \code{qswat_project} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print QSWATPlus Project Summary
}
