% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{qswat_plot_hrus}
\alias{qswat_plot_hrus}
\title{Plot HRU Map}
\usage{
qswat_plot_hrus(project, color_by = "landuse", title = "HRUs")
}
\arguments{
\item{project}{A \code{qswat_project} object.}

\item{color_by}{Character. Column in \code{project$hru_sf} used to colour the
polygons. Typical choices: \code{"landuse"}, \code{"soil"}, \code{"slope_class"},
\code{"hru_id"}. Default \code{"landuse"}.}

\item{title}{Character. Map title.}
}
\value{
A \code{tmap} object.
}
\description{
Creates a spatial map of HRU polygons, coloured by a chosen attribute.
Requires \code{\link[=qswat_create_hrus]{qswat_create_hrus()}} to have been run.
}
\examples{
\dontrun{
project <- qswat_create_hrus(project, lu_lookup, soil_lookup)
qswat_plot_hrus(project)
qswat_plot_hrus(project, color_by = "soil")
}

}
